requirements.txt
streamlit
streamlit_paste_button
google.generativeai 